pyproject.toml
[tool.poetry]
name = "djangoproject2"
version = "0.1.0"
description = ""
authors = ["sattorovsharofiddin <cyberpancher10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
django-filter = "^23.2"
coreapi = "^2.3.3"
pypdf2 = "^3.0.1"
pdf2image = "^1.16.3"
djangorestframework-simplejwt = "^5.2.2"
pypdf4 = "^1.27.0"
celery = "^5.3.1"
pdfplumber = "^0.9.0"
django-celery-results = "^2.5.1"
redis = "^4.6.0"
boto3 = "^1.28.1"
django-minio-backend = "3.5.0"
django-storages = "1.13.2"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
